// CRKing7_FE/src/types/address.type.ts
export interface Address {
  addressDetail: string;
  createdDate: string;
  district: string;
  firstName: string;
  focus: number;
  id: number;
  lastName: string;
  modifiedDate: string;
  phone: string;
  province: string;
  status: number;
  wards: string;
}
// const bb = {
//   addressDetail
// : 
// "Nguyệt đức"
// createdDate
// : 
// "2023-09-05T14:00:01.996+07:00"
// district
// : 
// "42"
// firstName
// : 
// "Kim Thăng"
// focus
// : 
// 0
// id
// : 
// 1
// lastName
// : 
// "Nguyễn"
// modifiedDate
// : 
// "2023-09-05T14:00:01.996+07:00"
// phone
// : 
// "0966821574"
// province
// : 
// "4"
// status
// : 
// 1
// wards
// : 
// "1297"
// }